Extract humanize CLI commands
